chore: improve code quality with gosec G115 overflow checks (#12)
- Re-enable gosec linter with G115 integer overflow checks
- Add nilerr and unconvert linters
- Add SafeInt32 and SafeInt32FromValue helper functions in state_helpers.go
  for safe int64 to int32 conversion with overflow checking
- Update all datasources and resources to use SafeInt32 helpers instead of
  direct int32() casts
- Remove unnecessary type conversions flagged by unconvert linter
- Add #nosec comments to test files where values are known to be safe
  (controlled test data, API-generated IDs)
- All acceptance tests pass against local Netbox container

+++ internal/resources/circuit_group_assignment_resource.go
@@ -205,7 +205,7 @@ func (r *CircuitGroupAssignmentResource) Read(ctx context.Context, req resource.
 	})
 
 	// Call the API to read the circuit group assignment
-	assignment, httpResp, err := r.client.CircuitsAPI.CircuitsCircuitGroupAssignmentsRetrieve(ctx, int32(id)).Execute()
+	assignment, httpResp, err := r.client.CircuitsAPI.CircuitsCircuitGroupAssignmentsRetrieve(ctx, id).Execute()
 
 	if err != nil {
 		// Check if resource was deleted outside of Terraform
@@ -304,7 +304,7 @@ func (r *CircuitGroupAssignmentResource) Update(ctx context.Context, req resourc
 	}
 
 	// Call the API to update the circuit group assignment
-	assignment, httpResp, err := r.client.CircuitsAPI.CircuitsCircuitGroupAssignmentsUpdate(ctx, int32(id)).
+	assignment, httpResp, err := r.client.CircuitsAPI.CircuitsCircuitGroupAssignmentsUpdate(ctx, id).
 		WritableCircuitGroupAssignmentRequest(*assignmentRequest).
 		Execute()
 
@@ -349,7 +349,7 @@ func (r *CircuitGroupAssignmentResource) Delete(ctx context.Context, req resourc
 	})
 
 	// Call the API to delete the circuit group assignment
-	httpResp, err := r.client.CircuitsAPI.CircuitsCircuitGroupAssignmentsDestroy(ctx, int32(id)).Execute()
+	httpResp, err := r.client.CircuitsAPI.CircuitsCircuitGroupAssignmentsDestroy(ctx, id).Execute()
 
 	if err != nil {
 		// Ignore 404 errors (resource already deleted)
@@ -386,7 +386,7 @@ func (r *CircuitGroupAssignmentResource) ImportState(ctx context.Context, req re
 	})
 
 	// Call the API to read the circuit group assignment
-	assignment, httpResp, err := r.client.CircuitsAPI.CircuitsCircuitGroupAssignmentsRetrieve(ctx, int32(id)).Execute()
+	assignment, httpResp, err := r.client.CircuitsAPI.CircuitsCircuitGroupAssignmentsRetrieve(ctx, id).Execute()
 
 	if err != nil {
 		resp.Diagnostics.AddError(
